Initial Draft - Project 1 Final

+++ controllers/admin.js
@@ -1,11 +1,16 @@
+const { validationResult } = require('express-validator');
+
 const Product = require('../models/product');
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
     pageTitle: 'Add Product',
     path: '/admin/add-product',
+    hasError: false,
+    errorMessage: undefined,
     // added globally
     //isAuthenticated: req.session.isLoggedIn,
+    validationErrors: [],
     editing: false
   });
 };
@@ -21,6 +26,28 @@ exports.postAddProduct = (req, res, next) => {
         tag = tag.trim();
       return tag;
     }) : [];
+  
+  const errors = validationResult(req);
+  // check for errors
+  if (!errors.isEmpty()) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      // added globally
+      //isAuthenticated: req.session.isLoggedIn,
+      editing: false,
+      hasError: true,
+      errorMessage: errors.array()[0].msg,
+      product: {
+        title: title,
+        imageUrl: imageUrl,
+        price: price,
+        description: description,
+        tags: tags,
+      },
+      validationErrors: errors.array(),
+    });
+  }
 
   // trim input as needed. Ideally, this would be done on the browser
   console.log("Here are the tags:", tags);
@@ -42,6 +69,33 @@ exports.postAddProduct = (req, res, next) => {
     })
     .catch(err => {
       console.log(err);
+      // for now, use Express to render the error page
+      const error = new Error(err); // 'Creating a new product failed.');
+      error.httpStatusCode = 500;
+      // throw the error onto until an error-handling middleware catches it
+      return next(error);
+      // we could redirect with a 500 error code
+      //res.status(500).redirect('/500'); 
+
+      // another option
+      // return res.status(500).render('admin/edit-product', {
+      //   pageTitle: 'Add Product',
+      //   path: '/admin/add-product',
+      //   // added globally
+      //   //isAuthenticated: req.session.isLoggedIn,
+      //   editing: false,
+      //   hasError: true,
+      //   errorMessage: "Database operation failed. Please try again.",
+      //   product: {
+      //     title: updatedTitle,
+      //     imageUrl: updatedImageUrl,
+      //     price: updatedPrice,
+      //     description: updatedDescription,
+      //     tags: updatedTags,
+      //     _id: prodId,
+      //   },
+      //   validationErrors: [],
+      // });
     });
 
 };
@@ -63,15 +117,26 @@ exports.getEditProduct = (req, res, next) => {
       if (!product) {
         return res.redirect('/');
       }
+      // the default status code for render is 200
       res.render('admin/edit-product', {
         pageTitle: 'Edit Product',
         path: '/admin/edit-product',
+        errorMessage: undefined,
+        hasError: false,
         editing: editMode,
         //isAuthenticated: req.session.isLoggedIn,
-        product: product
+        product: product,
+        validationErrors: [],
       });
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+      console.log(err)
+      // for now, use middleware to render the error page
+      const error = new Error(err); // 'Creating a new product failed.');
+      error.httpStatusCode = 500;
+      // throw the error onto until an error-handling middleware catches it
+      return next(error);
+    });
 };
 
 exports.postEditProduct = (req, res, next) => {
@@ -88,8 +153,35 @@ exports.postEditProduct = (req, res, next) => {
       return tag;
     }) : [];
 
+  const errors = validationResult(req);
+  // check for errors
+  if (!errors.isEmpty()) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Edit Product',
+      path: '/admin/edit-product',
+      // added globally
+      //isAuthenticated: req.session.isLoggedIn,
+      editing: true,
+      hasError: true,
+      errorMessage: errors.array()[0].msg,
+      product: {
+        title: updatedTitle,
+        imageUrl: updatedImageUrl,
+        price: updatedPrice,
+        description: updatedDescription,
+        tags: updatedTags,
+        _id: prodId,
+      },
+      validationErrors: errors.array(),
+    });
+  }
+
   Product.findById(prodId)
     .then(product => {
+      if (product.userId.toString() !== req.user._id.toString()) {
+        // the default status code for redirect is 300
+        return res.redirect('/');
+      }
       // mongoose object
       product.title = updatedTitle;
       product.price = updatedPrice;
@@ -98,11 +190,42 @@ exports.postEditProduct = (req, res, next) => {
       product.tags = updatedTags;
       console.log("Saving the edited product:", product);
       product
-        .save() // does an update
-      console.log("Updated Product");
-      res.redirect('/admin/products');
+        .save().then(result => {
+          console.log("Updated Product");
+          res.redirect('/admin/products');
+        }) // does an update
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+      console.log(err);
+      // for now, use Express to render the error page
+      const error = new Error(err); // 'Creating a new product failed.');
+      error.httpStatusCode = 500;
+      // throw the error onto until an error-handling middleware catches it
+      return next(error);
+
+      // we could  redirect with a 500 error code
+      //res.status(500).redirect('/500'); 
+      // we could just update the currently-used view
+      // return res.status(500).render('admin/edit-product', {
+      //   pageTitle: 'Edit Product',
+      //   path: '/admin/edit-product',
+      //   // added globally
+      //   //isAuthenticated: req.session.isLoggedIn,
+      //   editing: true,
+      //   hasError: true,
+      //   errorMessage: "Database operation failed. Please try again.",
+      //   product: {
+      //     title: updatedTitle,
+      //     imageUrl: updatedImageUrl,
+      //     price: updatedPrice,
+      //     description: updatedDescription,
+      //     tags: updatedTags,
+      //     _id: prodId,
+      //   },
+      //   validationErrors: [],
+      // });
+
+    });
   // const updatedProduct = new Product({
   //   updatedTitle,
   //   updatedImageUrl,
@@ -125,7 +248,10 @@ exports.postEditProduct = (req, res, next) => {
 
 
 exports.getProducts = (req, res, next) => {
-  Product.find()
+  // filter by the currently logged in user
+  // so that not all products are shown
+  /// req.user came from previous middleware
+  Product.find({userId: req.user._id})
     // populate populates a field with all the information and not just the id - Maxamillion
     // .populate('userId', 'name')
     // select allows you to specify which fields should be returned from the database
@@ -134,12 +260,19 @@ exports.getProducts = (req, res, next) => {
       res.render('admin/products', {
         prods: products,
         pageTitle: "Admin Products",
+        errorMessage: undefined,
         //isAuthenticated: req.session.isLoggedIn,
-        path: '/admin/products'
+        path: '/admin/products',
+        validationErrors: [],
       });
     })
     .catch(err => {
       console.log(err);
+      // for now, use Express to render the error page
+      const error = new Error(err); // 'Creating a new product failed.');
+      error.httpStatusCode = 500;
+      // throw the error onto until an error-handling middleware catches it
+      return next(error);
     });
   // Product.fetchAll(products => {
   //   res.render('admin/products', {
@@ -152,13 +285,18 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.findByIdAndRemove(prodId)
-    .then(() => {
+  Product.deleteOne({ _id: prodId, userId: req.user._id })//.findByIdAndRemove(prodId)
+    .then(result => {
       console.log("Product deleted");
       res.redirect('/admin/products');
     })
     .catch(err => {
       console.log(err);
+      // for now, use middleware to render the error page
+      const error = new Error(err); // 'Creating a new product failed.');
+      error.httpStatusCode = 500;
+      // throw the error onto until an error-handling middleware catches it
+      return next(error);
     });
 };
 
@@ -176,6 +314,32 @@ exports.getUserPastSearchHistory = (req, res, next) => {
     res.setHeader('Content-Type', 'application/json');
     res.end(JSON.stringify([])); // no history
   }
+};
+
+// 'admin/user/authorizations' => GET
+exports.getUserAuthorizations = (req, res, next) => {
+  if (req.user) {
+    return res.render('admin/user-authorizations', {
+      path: '/admin/user/authorizations',
+      pageTitle: 'Admin User Authorizations',
+      users: [req.user],
+    }, function(err, html) {
+      // error handling flow from https://stackoverflow.com/questions/7283896/how-can-i-catch-a-rendering-error-missing-template-in-node-js-using-express-js/15689798
+      if (err) {
+        console.log(err)
+        res.redirect('/404'); // File doesn't exist
+      } else {
+        res.send(html);
+      }
+    });
+  }
+  // else redirect to the login screen
+  res.redirect('/login')
+};
+
+// 'admin/user/authorizations' => POST
+exports.postUserAuthorizations = (req, res, next) => {
+  res.redirect('/');
 };
 
 // 'admin/update-ids' => POST
